Added camera rotation and scaled up the size of scene02 and 03

+++ DM2122_Framework/Application/Source/Camera.cpp
@@ -61,18 +61,18 @@ void Camera::Update(double dt)
 {
 	if (Application::IsKeyPressed('W'))
 	{
-		position.y += 10.0f *dt;
+		position.y += 10.0f * (float)(dt);
 	}
 	if (Application::IsKeyPressed('S'))
 	{
-		position.y -= 10.0f *dt;
+		position.y -= 10.0f * (float)(dt);
 	}
 	if (Application::IsKeyPressed('A'))
 	{
-		position.x -= 10.0f *dt;
+		position.x -= 10.0f * (float)(dt);
 	}
 	if (Application::IsKeyPressed('D'))
 	{
-		position.x += 10.0f *dt;
+		position.x += 10.0f * (float)(dt);
 	}
 }
